+++ locales/enUS.lua
@@ -1,4 +1,4 @@
-local TranqRotate = select(2, ...)
+TranqRotate = select(2, ...)
 
 local L = {
 
@@ -45,6 +45,7 @@ local L = {
     ["SETTING_ANNOUNCES"] = "Announces",
     ["ENABLE_ANNOUNCES"] = "Enable announces",
     ["ENABLE_ANNOUNCES_DESC"] = "Enable / disable the announcement.",
+    ["YELL_SAY_DISABLED_OPEN_WORLD"] = "(Yell and say channels does not work in open world, but will inside your raids)",
 
     ---- Channels
     ["ANNOUNCES_CHANNEL_HEADER"] = "Announce channel",