+++ src/rotation.lua
@@ -1,5 +1,3 @@
-local TranqRotate = select(2, ...)
-
 local L = TranqRotate.L
 
 -- Adds hunter to global table and one of the two rotation tables
@@ -12,6 +10,8 @@ function TranqRotate:registerHunter(hunterName)
     hunter.frame = nil
     hunter.nextTranq = false
     hunter.lastTranqTime = 0
+    hunter.lastFailTime = 0
+    hunter.addonVersion = nil
 
     -- Add to global list
     table.insert(TranqRotate.hunterTable, hunter)
@@ -53,19 +53,11 @@ function TranqRotate:removeHunter(deletedHunter)
 end
 
 -- Update the rotation list once a tranq has been done.
--- The parameter is the hunter that used it's tranq (successfully or not)
-function TranqRotate:rotate(lastHunter, fail, rotateWithoutCooldown)
+-- The parameter is the hunter that used it's tranq
+function TranqRotate:rotate(lastHunter, rotateWithoutCooldown)
 
-    -- Default value to false
-    fail = fail or false
-
-    if (not fail) then
-        TranqRotate.frenzy = false
-    end
-
-    local playerName, realm = UnitName("player")
+    TranqRotate.frenzy = false
     local lastHunterRotationTable = TranqRotate:getHunterRotationTable(lastHunter)
-    local hasPlayerFailed = playerName == lastHunter.name and fail
 
     lastHunter.lastTranqTime = GetTime()
 
@@ -81,25 +73,52 @@ function TranqRotate:rotate(lastHunter, fail, rotateWithoutCooldown)
 
         if (nextHunter ~= nil) then
             TranqRotate:setNextTranq(nextHunter)
-
-            if ((fail and nextHunter.name == playerName) and
-                #TranqRotate.rotationTables.backup < 1 and
-                TranqRotate:isHunterTranqCooldownReady(nextHunter)
-            ) then
-                TranqRotate:throwTranqAlert()
-            end
         end
     end
+end
 
-    if (fail) then
-        if (hasPlayerFailed) then
-            TranqRotate:alertBackup(TranqRotate.db.profile.whisperFailMessage, nextHunter, true)
-        end
+-- Handle miss or dispel resist scenario
+function TranqRotate:handleFailTranq(hunter, event)
 
-        local playerRotationTable = TranqRotate:getHunterRotationTable(TranqRotate:getHunter(playerName))
-        if (playerRotationTable == TranqRotate.rotationTables.backup and not hasPlayerFailed) then
-            TranqRotate:throwTranqAlert()
-        end
+    -- Do not process multiple SPELL_DISPEL_FAILED events or multiple fail broadcasts
+    local duplicate = hunter.lastFailTime >=  GetTime() - TranqRotate.constants.duplicateFailedTranqshotDelayThreshold
+    if (duplicate) then
+        return
+    end
+
+    TranqRotate:printFail(hunter, event)
+
+    local playerName, realm = UnitName("player")
+    local hasPlayerFailed = playerName == hunter.name
+    local nextHunter = TranqRotate:getHighlightedHunter()
+    local lastHunterRotationTable = TranqRotate:getHunterRotationTable(hunter)
+
+    -- Could happen if the first event received is a miss/resist
+    if (nextHunter == nil) then
+        nextHunter = TranqRotate:getNextRotationHunter(hunter)
+    end
+
+    hunter.lastFailTime = GetTime()
+
+    -- No backup, if player is next in rotation he will be warned to handle the fail
+    if (
+        lastHunterRotationTable == TranqRotate.rotationTables.rotation and
+        nextHunter.name == playerName and
+        #TranqRotate.rotationTables.backup < 1 and
+        TranqRotate:isHunterTranqCooldownReady(nextHunter)
+    ) then
+        TranqRotate:throwTranqAlert()
+    end
+
+    -- The player failed, sending fail message and backup alerts
+    if (hasPlayerFailed) then
+        TranqRotate:alertBackup(TranqRotate.db.profile.whisperFailMessage, nextHunter, true)
+    end
+
+    -- Player is in backup group, display an alert when someone fails
+    local playerRotationTable = TranqRotate:getHunterRotationTable(TranqRotate:getHunter(playerName))
+    if (playerRotationTable == TranqRotate.rotationTables.backup and not hasPlayerFailed) then
+        TranqRotate:throwTranqAlert()
     end
 end
 
@@ -215,12 +234,20 @@ end
 -- @todo: remove this | TEST FUNCTION - Manually rotate hunters for test purpose
 function TranqRotate:testRotation()
 
+    local hunterToRotate = nil
     for key, hunter in pairs(TranqRotate.rotationTables.rotation) do
         if (hunter.nextTranq) then
-            TranqRotate:rotate(hunter, false)
+            hunterToRotate = hunter
             break
         end
     end
+
+    if (not hunterToRotate) then
+        hunterToRotate = TranqRotate.rotationTables.rotation[1]
+    end
+
+    TranqRotate:sendSyncTranq(hunterToRotate, false, timestamp)
+    TranqRotate:rotate(hunterToRotate)
 end
 
 -- Check if a hunter is already registered
@@ -327,7 +354,7 @@ function TranqRotate:updateHunterStatus(hunter)
 
     -- Jump to the next hunter if the current one is dead or offline
     if (hunter.nextTranq and (not TranqRotate:isHunterAliveAndOnline(hunter))) then
-        TranqRotate:rotate(hunter, false, true)
+        TranqRotate:rotate(hunter, true)
     end
 
     TranqRotate:refreshHunterFrame(hunter)
@@ -465,12 +492,14 @@ function TranqRotate:alertBackup(message, nextHunter, noComms)
             nextHunter = TranqRotate:getNextRotationHunter(player)
         end
 
-        SendChatMessage(message, 'WHISPER', nil, nextHunter.name)
-        if (noComms ~= true) then
-            TranqRotate:sendBackupRequest(nextHunter.name)
+        if (playerName ~= nextHunter.name) then
+            SendChatMessage(message, 'WHISPER', nil, nextHunter.name)
+            if (noComms ~= true) then
+                TranqRotate:sendBackupRequest(nextHunter.name)
+            end
         end
     else
-        TranqRotate:whisperBackup(message)
+        TranqRotate:whisperBackup(message, noComms)
     end
 end
 
@@ -492,4 +521,16 @@ function TranqRotate:whisperBackup(message, noComms)
             end
         end
     end
+end
+
+-- Returns the hunter currently wearing the "next" flag
+function TranqRotate:getHighlightedHunter()
+
+    for key,hunter in pairs(TranqRotate.rotationTables.rotation) do
+        if (hunter.nextTranq) then
+            return hunter
+        end
+    end
+
+    return nil
 end